+++ modules/apps/portal/portal-cluster-sample/bnd.bnd
@@ -1,3 +1,3 @@
-Bundle-Name: Liferay Cluster Test Module
-Bundle-SymbolicName: com.liferay.cluster.test.module
-Bundle-Version: 3.0.0
+Bundle-Name: Liferay Portal Cluster Sample
+Bundle-SymbolicName: com.liferay.portal.cluster.sample
+Bundle-Version: 1.0.0
