+++ modules/apps/portal/portal-cluster-sample/src/main/java/com/liferay/portal/cluster/sample/internal/ClusterTestCommands.java
@@ -12,33 +12,21 @@
  * details.
  */
 
-package com.liferay.cluster.test.module.internel;
-
-import com.liferay.portal.kernel.cluster.ClusterExecutorUtil;
-import com.liferay.portal.kernel.cluster.ClusterNode;
-
-import java.io.Serializable;
+package com.liferay.portal.cluster.sample.internal;
 
 /**
  * @author Tina Tian
  */
-public class ClusterTestClass implements Serializable {
+public class ClusterTestCommands {
 
-	public static int getPortalLocalPort() {
-		ClusterNode clusterNode = ClusterExecutorUtil.getLocalClusterNode();
+	public static final String INVOKE_METHOD_MODULE = "invoke-method-module";
 
-		return clusterNode.getPortalPort();
-	}
+	public static final String INVOKE_METHOD_MODULE_ON_MASTER =
+		"invoke-method-module-on-master";
 
-	public ClusterTestClass(String name) {
-		_name = name;
-	}
+	public static final String INVOKE_METHOD_PORTAL = "invoke-method-portal";
 
-	@Override
-	public String toString() {
-		return _name;
-	}
-
-	private final String _name;
+	public static final String INVOKE_METHOD_PORTAL_ON_MASTER =
+		"invoke-method-portal-on-master";
 
 }
