+++ modules/apps/portal/portal-cluster-sample/src/main/resources/META-INF/module-log4j.xml
